Reapply "Refactor to move ChartState to vegafusion core (#519)" (#520)
This reverts commit 713b1597c713aa31a83b067cb480dd0eab08d045.
