+++ Cargo.toml
@@ -31,6 +31,7 @@ prost = { version = "0.12.3" }
 prost-types = { version = "0.12.3" }
 object_store = { version = "0.11.0" }
 lazy_static = { version = "1.5" }
+async-trait = "0.1.73"
 
 [workspace.dependencies.datafusion]
 version = "42.0.0"
