+++ vegafusion-core/Cargo.toml
@@ -33,6 +33,12 @@ workspace = true
 [dependencies.prost-types]
 workspace = true
 
+[dependencies.chrono-tz]
+workspace = true
+
+[dependencies.async-trait]
+workspace = true
+
 [dependencies.sqlparser]
 workspace = true
 optional = true
@@ -48,6 +54,10 @@ version = "1.6.9"
 
 [dependencies.datafusion-common]
 workspace = true
+
+[dependencies.vegafusion-dataframe]
+path = "../vegafusion-dataframe"
+version = "1.6.9"
 
 [dependencies.pyo3]
 workspace = true
