+++ vegafusion-core/src/data/dataset.rs
@@ -1,8 +1,8 @@
+use crate::error::Result;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 use vegafusion_common::data::table::VegaFusionTable;
-use vegafusion_core::error::Result;
 use vegafusion_dataframe::dataframe::DataFrame;
 
 #[derive(Clone)]
