+++ vegafusion-core/src/data/mod.rs
@@ -1,1 +1,2 @@
+pub mod dataset;
 pub mod tasks;
