+++ vegafusion-core/src/lib.rs
@@ -2,11 +2,13 @@
 extern crate lazy_static;
 extern crate core;
 
+pub mod chart_state;
 pub mod data;
 pub mod expression;
 pub mod patch;
 pub mod planning;
 pub mod proto;
+pub mod runtime;
 pub mod spec;
 pub mod task_graph;
 pub mod transform;
