+++ vegafusion-datafusion-udfs/Cargo.toml
@@ -22,6 +22,10 @@ workspace = true
 path = "../vegafusion-common"
 version = "1.6.9"
 
+[dependencies.vegafusion-core]
+path = "../vegafusion-core"
+version = "1.6.9"
+
 [dependencies.datafusion-physical-expr]
 workspace = true
