+++ vegafusion-python/Cargo.toml
@@ -15,9 +15,11 @@ protobuf-src = ["vegafusion-core/protobuf-src"]
 [dependencies]
 log = "0.4.17"
 env_logger = "0.10.0"
-async-trait = "0.1.73"
 uuid = "1.3.0"
 sysinfo = "0.32.0"
+
+[dependencies.async-trait]
+workspace = true
 
 [dependencies.lazy_static]
 workspace = true
