+++ vegafusion-runtime/src/data/mod.rs
@@ -1,2 +1,1 @@
-pub mod dataset;
 pub mod tasks;
