+++ vegafusion-runtime/src/lib.rs
@@ -4,7 +4,6 @@ extern crate core;
 
 pub mod data;
 pub mod expression;
-pub mod pre_transform;
 pub mod signal;
 pub mod task_graph;
 pub mod tokio_runtime;
