--- vegafusion-runtime/src/pre_transform/mod.rs
@@ -1,1 +1,0 @@
-pub mod destringify_selection_datetimes;
