+++ vegafusion-runtime/src/task_graph/runtime.rs
@@ -1,30 +1,22 @@
-use async_recursion::async_recursion;
-use std::collections::{HashMap, HashSet};
-use vegafusion_core::error::{Result, ResultWithContext, ToExternalError, VegaFusionError};
-use vegafusion_core::task_graph::task_value::TaskValue;
-
-use crate::data::dataset::VegaFusionDataset;
-use crate::pre_transform::destringify_selection_datetimes::destringify_selection_datetimes;
 use crate::task_graph::cache::VegaFusionCache;
 use crate::task_graph::task::TaskCall;
 use crate::task_graph::timezone::RuntimeTzConfig;
-use datafusion_common::ScalarValue;
+use async_recursion::async_recursion;
 use futures_util::{future, FutureExt};
 use prost::Message as ProstMessage;
-use serde_json::Value;
+use std::collections::HashMap;
 use std::convert::{TryFrom, TryInto};
 use std::panic::AssertUnwindSafe;
-use std::sync::{Arc, Mutex};
-use vegafusion_common::data::scalar::ScalarValueHelpers;
+use std::sync::Arc;
 use vegafusion_common::data::table::VegaFusionTable;
+use vegafusion_core::data::dataset::VegaFusionDataset;
+use vegafusion_core::error::{Result, ResultWithContext, ToExternalError, VegaFusionError};
+use vegafusion_core::planning::apply_pre_transform::apply_pre_transform_datasets;
+use vegafusion_core::planning::destringify_selection_datetimes::destringify_selection_datetimes;
 use vegafusion_core::planning::plan::{PlannerConfig, SpecPlan};
-use vegafusion_core::planning::stitch::CommPlan;
-use vegafusion_core::planning::watch::{
-    ExportUpdateArrow, ExportUpdateJSON, ExportUpdateNamespace,
-};
+use vegafusion_core::planning::watch::{ExportUpdateArrow, ExportUpdateNamespace};
 use vegafusion_core::proto::gen::errors::error::Errorkind;
 use vegafusion_core::proto::gen::errors::{Error, TaskGraphValueError};
-use vegafusion_core::proto::gen::pretransform::pre_transform_spec_warning::WarningType;
 use vegafusion_core::proto::gen::pretransform::pre_transform_values_warning::WarningType as ValuesWarningType;
 use vegafusion_core::proto::gen::pretransform::{
     pre_transform_extract_warning, PlannerWarning,
@@ -33,8 +25,7 @@ use vegafusion_core::proto::gen::pretransform::{
     PreTransformValuesRequest, PreTransformValuesResponse, PreTransformValuesWarning,
 };
 use vegafusion_core::proto::gen::pretransform::{
-    PreTransformBrokenInteractivityWarning, PreTransformRowLimitWarning, PreTransformSpecRequest,
-    PreTransformSpecResponse, PreTransformUnsupportedWarning,
+    PreTransformRowLimitWarning, PreTransformSpecRequest, PreTransformSpecResponse,
 };
 use vegafusion_core::proto::gen::services::{
     pre_transform_extract_result, pre_transform_spec_result, pre_transform_values_result,
@@ -43,11 +34,13 @@ use vegafusion_core::proto::gen::services::{
 };
 use vegafusion_core::proto::gen::tasks::{
     task::TaskKind, InlineDataset, NodeValueIndex, ResponseTaskValue, TaskGraph,
-    TaskGraphValueResponse, TaskValue as ProtoTaskValue, TzConfig, Variable, VariableNamespace,
+    TaskGraphValueResponse, TaskValue as ProtoTaskValue, TzConfig, VariableNamespace,
 };
+use vegafusion_core::runtime::VegaFusionRuntimeTrait;
 use vegafusion_core::spec::chart::ChartSpec;
 use vegafusion_core::spec::values::MissingNullOrValue;
 use vegafusion_core::task_graph::graph::ScopedVariable;
+use vegafusion_core::task_graph::task_value::TaskValue;
 use vegafusion_dataframe::connection::Connection;
 
 type CacheValue = (TaskValue, Vec<TaskValue>);
@@ -716,6 +709,60 @@ impl VegaFusionRuntime {
     }
 }
 
+#[async_trait::async_trait]
+impl VegaFusionRuntimeTrait for VegaFusionRuntime {
+    async fn query_request(
+        &self,
+        task_graph: Arc<TaskGraph>,
+        indices: &[NodeValueIndex],
+        inline_datasets: &HashMap<String, VegaFusionDataset>,
+    ) -> Result<Vec<ResponseTaskValue>> {
+        // Clone task_graph and task_graph_runtime for use in closure
+        let task_graph_runtime = self.clone();
+        let response_value_futures: Vec<_> = indices
+            .iter()
+            .map(|node_value_index| {
+                let node = task_graph
+                    .nodes
+                    .get(node_value_index.node_index as usize)
+                    .with_context(|| {
+                        format!(
+                            "Node index {} out of bounds for graph with size {}",
+                            node_value_index.node_index,
+                            task_graph.nodes.len()
+                        )
+                    })?;
+                let task = node.task();
+                let var = match node_value_index.output_index {
+                    None => task.variable().clone(),
+                    Some(output_index) => task.output_vars()[output_index as usize].clone(),
+                };
+
+                let scope = node.task().scope.clone();
+
+                // Clone task_graph and task_graph_runtime for use in closure
+                let task_graph_runtime = task_graph_runtime.clone();
+                let task_graph = task_graph.clone();
+
+                Ok(async move {
+                    let value = task_graph_runtime
+                        .clone()
+                        .get_node_value(task_graph, node_value_index, inline_datasets.clone())
+                        .await?;
+
+                    Ok::<_, VegaFusionError>(ResponseTaskValue {
+                        variable: Some(var),
+                        scope,
+                        value: Some(ProtoTaskValue::try_from(&value).unwrap()),
+                    })
+                })
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        future::try_join_all(response_value_futures).await
+    }
+}
+
 #[async_recursion]
 async fn get_or_compute_node_value(
     task_graph: Arc<TaskGraph>,
@@ -788,315 +835,5 @@ async fn get_or_compute_node_value(
         // get or construct from cache
 
         cache.get_or_try_insert_with(cache_key, fut).await
-    }
-}
-
-fn apply_pre_transform_datasets(
-    input_spec: &ChartSpec,
-    plan: &SpecPlan,
-    init: Vec<ExportUpdateArrow>,
-    row_limit: Option<u32>,
-) -> Result<(ChartSpec, Vec<PreTransformSpecWarning>)> {
-    // Update client spec with server values
-    let mut spec = plan.client_spec.clone();
-    let mut limited_datasets: Vec<Variable> = Vec::new();
-    for export_update in init {
-        let scope = export_update.scope.clone();
-        let name = export_update.name.as_str();
-        let export_update = export_update.to_json()?;
-        match export_update.namespace {
-            ExportUpdateNamespace::Signal => {
-                let signal = spec.get_nested_signal_mut(&scope, name)?;
-                signal.value = MissingNullOrValue::Value(export_update.value);
-            }
-            ExportUpdateNamespace::Data => {
-                let data = spec.get_nested_data_mut(&scope, name)?;
-
-                // If the input dataset includes inline values and no transforms,
-                // copy the input JSON directly to avoid the case where round-tripping
-                // through Arrow homogenizes mixed type arrays.
-                // E.g. round tripping may turn [1, "two"] into ["1", "two"]
-                let input_values = input_spec
-                    .get_nested_data(&scope, name)
-                    .ok()
-                    .and_then(|data| {
-                        if data.transform.is_empty() {
-                            data.values.clone()
-                        } else {
-                            None
-                        }
-                    });
-                let value = if let Some(input_values) = input_values {
-                    input_values
-                } else if let Value::Array(values) = &export_update.value {
-                    if let Some(row_limit) = row_limit {
-                        let row_limit = row_limit as usize;
-                        if values.len() > row_limit {
-                            limited_datasets.push(export_update.to_scoped_var().0);
-                            Value::Array(Vec::from(&values[..row_limit]))
-                        } else {
-                            Value::Array(values.clone())
-                        }
-                    } else {
-                        Value::Array(values.clone())
-                    }
-                } else {
-                    return Err(VegaFusionError::internal(
-                        "Expected Data value to be an Array",
-                    ));
-                };
-
-                // Set inline value
-                // Other properties are cleared by planning process so we don't alter them here
-                data.values = Some(value);
-            }
-        }
-    }
-
-    // Destringify datetime strings in selection store datasets
-    destringify_selection_datetimes(&mut spec)?;
-
-    // Build warnings
-    let mut warnings: Vec<PreTransformSpecWarning> = Vec::new();
-
-    // Add unsupported warning (
-    if plan.comm_plan.server_to_client.is_empty() {
-        warnings.push(PreTransformSpecWarning {
-            warning_type: Some(WarningType::Unsupported(PreTransformUnsupportedWarning {})),
-        });
-    }
-
-    // Add Row Limit warning
-    if !limited_datasets.is_empty() {
-        warnings.push(PreTransformSpecWarning {
-            warning_type: Some(WarningType::RowLimit(PreTransformRowLimitWarning {
-                datasets: limited_datasets,
-            })),
-        });
-    }
-
-    // Add Broken Interactivity warning
-    if !plan.comm_plan.client_to_server.is_empty() {
-        let vars: Vec<_> = plan
-            .comm_plan
-            .client_to_server
-            .iter()
-            .map(|var| var.0.clone())
-            .collect();
-        warnings.push(PreTransformSpecWarning {
-            warning_type: Some(WarningType::BrokenInteractivity(
-                PreTransformBrokenInteractivityWarning { vars },
-            )),
-        });
-    }
-
-    // Add planner warnings
-    for planner_warning in &plan.warnings {
-        warnings.push(PreTransformSpecWarning {
-            warning_type: Some(WarningType::Planner(PlannerWarning {
-                message: planner_warning.message(),
-            })),
-        });
-    }
-
-    Ok((spec, warnings))
-}
-
-#[derive(Clone)]
-pub struct ChartState {
-    input_spec: ChartSpec,
-    transformed_spec: ChartSpec,
-    plan: SpecPlan,
-    inline_datasets: HashMap<String, VegaFusionDataset>,
-    task_graph: Arc<Mutex<TaskGraph>>,
-    task_graph_mapping: Arc<HashMap<ScopedVariable, NodeValueIndex>>,
-    server_to_client_value_indices: Arc<HashSet<NodeValueIndex>>,
-    warnings: Vec<PreTransformSpecWarning>,
-}
-
-impl ChartState {
-    pub async fn try_new(
-        runtime: &VegaFusionRuntime,
-        spec: ChartSpec,
-        inline_datasets: HashMap<String, VegaFusionDataset>,
-        tz_config: TzConfig,
-        row_limit: Option<u32>,
-    ) -> Result<Self> {
-        let dataset_fingerprints = inline_datasets
-            .iter()
-            .map(|(k, ds)| (k.clone(), ds.fingerprint()))
-            .collect::<HashMap<_, _>>();
-
-        let plan = SpecPlan::try_new(&spec, &Default::default())?;
-
-        let task_scope = plan
-            .server_spec
-            .to_task_scope()
-            .with_context(|| "Failed to create task scope for server spec")?;
-        let tasks = plan
-            .server_spec
-            .to_tasks(&tz_config, &dataset_fingerprints)
-            .unwrap();
-        let task_graph = TaskGraph::new(tasks, &task_scope).unwrap();
-        let task_graph_mapping = task_graph.build_mapping();
-        let server_to_client_value_indices: Arc<HashSet<_>> = Arc::new(
-            plan.comm_plan
-                .server_to_client
-                .iter()
-                .map(|scoped_var| task_graph_mapping.get(scoped_var).unwrap().clone())
-                .collect(),
-        );
-
-        // Gather values of server-to-client values
-        let mut init = Vec::new();
-        for var in &plan.comm_plan.server_to_client {
-            let node_index = task_graph_mapping
-                .get(var)
-                .with_context(|| format!("Failed to lookup variable '{var:?}'"))?;
-            let value = runtime
-                .get_node_value(
-                    Arc::new(task_graph.clone()),
-                    node_index,
-                    inline_datasets.clone(),
-                )
-                .await
-                .with_context(|| "Failed to get node value")?;
-
-            init.push(ExportUpdateArrow {
-                namespace: ExportUpdateNamespace::try_from(var.0.namespace()).unwrap(),
-                name: var.0.name.clone(),
-                scope: var.1.clone(),
-                value,
-            });
-        }
-
-        let (transformed_spec, warnings) =
-            apply_pre_transform_datasets(&spec, &plan, init, row_limit)?;
-
-        Ok(Self {
-            input_spec: spec,
-            transformed_spec,
-            plan,
-            inline_datasets,
-            task_graph: Arc::new(Mutex::new(task_graph)),
-            task_graph_mapping: Arc::new(task_graph_mapping),
-            server_to_client_value_indices,
-            warnings,
-        })
-    }
-
-    pub async fn update(
-        &self,
-        runtime: &VegaFusionRuntime,
-        updates: Vec<ExportUpdateJSON>,
-    ) -> Result<Vec<ExportUpdateJSON>> {
-        let mut task_graph = self.task_graph.lock().map_err(|err| {
-            VegaFusionError::internal(format!("Failed to acquire task graph lock: {:?}", err))
-        })?;
-        let server_to_client = self.server_to_client_value_indices.clone();
-        let mut indices: Vec<NodeValueIndex> = Vec::new();
-        for export_update in &updates {
-            let var = match export_update.namespace {
-                ExportUpdateNamespace::Signal => Variable::new_signal(&export_update.name),
-                ExportUpdateNamespace::Data => Variable::new_data(&export_update.name),
-            };
-            let scoped_var: ScopedVariable = (var, export_update.scope.clone());
-            let node_value_index = self
-                .task_graph_mapping
-                .get(&scoped_var)
-                .with_context(|| format!("No task graph node found for {scoped_var:?}"))?
-                .clone();
-
-            let value = match export_update.namespace {
-                ExportUpdateNamespace::Signal => {
-                    TaskValue::Scalar(ScalarValue::from_json(&export_update.value)?)
-                }
-                ExportUpdateNamespace::Data => {
-                    TaskValue::Table(VegaFusionTable::from_json(&export_update.value)?)
-                }
-            };
-
-            indices.extend(task_graph.update_value(node_value_index.node_index as usize, value)?);
-        }
-
-        // Filter to update nodes in the comm plan
-        let indices: Vec<_> = indices
-            .iter()
-            .filter(|&node| server_to_client.contains(node))
-            .cloned()
-            .collect();
-
-        let cloned_task_graph = task_graph.clone();
-
-        // Drop the MutexGuard before await call to avoid warning
-        drop(task_graph);
-
-        let response_task_values = runtime
-            .query_request(
-                Arc::new(cloned_task_graph),
-                indices.as_slice(),
-                &self.inline_datasets,
-            )
-            .await?;
-
-        let mut response_updates = response_task_values
-            .into_iter()
-            .map(|response_value| {
-                let variable = response_value
-                    .variable
-                    .with_context(|| "Missing variable for response value".to_string())?;
-
-                let scope = response_value.scope;
-                let proto_value = response_value
-                    .value
-                    .with_context(|| "missing value for response value: {:?}".to_string())?;
-
-                let value = TaskValue::try_from(&proto_value).with_context(|| {
-                    "Deserialization failed for value of response value: {:?}".to_string()
-                })?;
-
-                Ok(ExportUpdateJSON {
-                    namespace: match variable.ns() {
-                        VariableNamespace::Signal => ExportUpdateNamespace::Signal,
-                        VariableNamespace::Data => ExportUpdateNamespace::Data,
-                        VariableNamespace::Scale => {
-                            return Err(VegaFusionError::internal("Unexpected scale variable"))
-                        }
-                    },
-                    name: variable.name.clone(),
-                    scope: scope.clone(),
-                    value: value.to_json()?,
-                })
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        // Sort for deterministic ordering
-        response_updates.sort_by_key(|update| update.name.clone());
-
-        Ok(response_updates)
-    }
-
-    pub fn get_input_spec(&self) -> &ChartSpec {
-        &self.input_spec
-    }
-
-    pub fn get_server_spec(&self) -> &ChartSpec {
-        &self.plan.server_spec
-    }
-
-    pub fn get_client_spec(&self) -> &ChartSpec {
-        &self.plan.client_spec
-    }
-
-    pub fn get_transformed_spec(&self) -> &ChartSpec {
-        &self.transformed_spec
-    }
-
-    pub fn get_comm_plan(&self) -> &CommPlan {
-        &self.plan.comm_plan
-    }
-
-    pub fn get_warnings(&self) -> &Vec<PreTransformSpecWarning> {
-        &self.warnings
     }
 }
